BUG: 174897
So that's it. It probably was a strange behavior caused by checking for XSync extension more than once,
even though I'm still convinced there is something wrong in X, since this behavior is way too strange...

Anyway, fixed now, making also the poller a singleton, so that we are sure initialization takes place only once.

svn path=/trunk/KDE/kdebase/workspace/; revision=892951