CCMAIL: ervin@kde.org
Adding a timeout feature to inhibition. This can be useful both in
case of cleaning up, both for some use cases. I have just added the
infrastructure as an overload, as I don't know what effect this could have
on Solid API. I'd use ::inhibit(QString,QString, int = <default>), but I
still fear this could break something. Kevin, I'd really like your opinion
on this, I think this feature is really needed and can be useful.

I am also planning to add a check through DBus, for that, wouldn't it be
smarter if we passed a KAboutData to inhibit? So we would have the application
friendly name for the notification and the service name we need for DBus.
I'd like your opinion on that too. We could deprecate the current inhibit, 
but my main concern is DBus and if we are actually forced to keep some kind of
API compatibility.


svn path=/trunk/KDE/kdebase/workspace/; revision=923178