BUG: 190011
Canceling pending suspend actions if the adaptor gets plugged in.
Tweaking the notification to let the user know what's going on.
Unfortunately, due to that I can't backport this, or better, I could
without letting the user know what's going on, and I don't know if this
is such a great idea


svn path=/trunk/KDE/kdebase/workspace/; revision=956038