BUG:203777
Ok, so in the end that crash was just the consequence of the fact that the initialization of the profiles
utterly failed in some way. I fixed that now, it just hung up for years here, but probably your bus was not
sustaining such a hard load.

Thanks for your report, this one was really a killer  ;)


svn path=/trunk/KDE/kdebase/workspace/; revision=1056565