Don't prematurely unload suspendsession action
Sometimes ActionPool::loadAction is called with an empty (invalid) KConfigGroup
in order to retrieve an Action* to trigger. These callers treat loadAction as a
glorified getter, and should not disturb a loaded action. However, an "active"
action was unloaded unconditionally and not subsequently reloaded, since
actions can only be loaded with valid configurations.

This meant that the idle timeout for the "SuspendSession" action was
unregistered when session suspend was triggered by PowerDevil::Core or by a
button event.

- Only unload action if it's going to be loaded (i.e. group isValid)
- Make m_activeActions track loaded actions, for consistency.

Note: only the ActionPool::loadAction("SuspendSession") callers suffer
undesirable behaviour from triggering this bug. The"DPMSControl" caller from
HandleButtonEvents::processAction would disable DPMS if another bug did not mask
this one.

BUG: 306950
REVIEW: 106692