powerdevil: Handle startup of UPower correctly.
It might occur that UPower is not correctly started by the activation service, for any reason.
Should that be the case, we of course should mark UPower as unavailable. This patch makes
the loading through activation safe, as we watch the service until it comes up, or timeout after 10 seconds.

Patch by me and Oliver.

BUG: 316472