Daemon: Use the correct method for setting global shortcuts
KGlobalAccel has a concept of "Deafult" and "Active" shortcuts. Default
shortcuts are just written in a config file, but are not used. The ones
that are used are the "Active" shortcuts.

We should either set both of them or use the new awesome function called
"setGlobalShortcut" which does all of that for us.