Don't show OSD when setting brightness absolutely
This way the OSD does not pop up when the user manipulates the brightness
from the battery monitor, it still does when using the brightness keys,
as has been discussed months ago on ml.

This should also "fix" other users of this interface, such as Kamoso, to
have the OSD pop up when it uses the screen as flashlight. Any other users
of this can still call the osdService manually if desired.

BUG: 336449