Don't show OSD when brightness changes in the backend
We only show the OSD when the user explicitly manipulates brighntess using the keys
which we handle too (since there's a convenient FDO dbus iface for that in contrast
to the screen stuff), so we can safely trigger the OSD from there.

This fixes battery monitor showing the OSD while dragging the keyboard brightness slider.