Do not start the fade effect if we are inhibit
When an inhibition for screen is added we modify some settings to
disable DPMS. The thing is we are not disabling the idleTimeout
registered in PowerDevilCore so idleTimeout is being called even
though we are inhibit.

We could remove the timeout when we get an inhibition (infact I believe
that would be a better fix) but I haven't been able to figure out how
to do it properly.

CCMAIL:kde@privat.broulik.de