Expose inhibitions over dbus
This adds a new "ListInhibitions" method and "InhibitionsChanged" signal to
the org.kde.Solid.PowerManagement.PolicyAgent object allowing others to know
who has posted an inhibition and why.

ListInhibitions returns a QList<QPair<QString, QString>> with the unprocessed
application name and the reason provided by the application.
InhibitionsChanged has a QList<QPair<QString, QString>> of the inhibitions
that were added and a QStringList of the applications that have removed their
inhibition. It does not do any formatting or duplicate removal.

REVIEW: 122130