Inhibitions posted over the freedesktop interface should always inhibit everything
When using Solid you can specify whether you want to prevent only suspending the session
(eg. copy job) or everything (eg. watching a movie). The standardized FDO interface does
not know this distinction and so we need to inhibit everything.

This fixes for instance the screen turning off while watching a video in Chrome (even Flash!)
but KDE applications like Amarok can let the screen turn off but the device stay on.