Change the final warning from 30 seconds to 1 minute
As decided on at the sprint having seen me run in panic up and down some
stairs to try and fetch and plug in my charger within the 30 second
warning.

This happened on numerous occasions and even with a team effort was
unable to reach my laptop in time before it suspended. One minute should
give people a chance resulting in fewer shutdowns and fewer injuries.

Reviewed by Kai Uwe Broulik