Change default low battery threshold to 10%
Back in the days when laptops ran for one hour that was reasonable; however
today 20% of battery easily tranlates to 1 hour of battery life, or more.
Thus when you get that notification you ignore it and when the critical battery
timer eventually kicks in you potentially stumble down the staircase rushing
for your AC adpater.