Add support for Display Data Control
This uses the ddccontrol utility to enable you to control the brightness of a desktop
monitor attached to your computer rather than your internal panel on a notebook.

It is the last item in the fallback chain: First it tries to use XRandR to adjust the
brightness, then it tries to use the SysFS helper thingie, and if all that fails, it
tries to use ddcontrol.