Simulate user activity when inhibitions change
This way when an inhibition outlasts an idle action (eg. 10 minute timeout for
auto-suspend but download takes 11 minutes) we make sure that 10 minutes after
the download has finished, the system suspends, as well as vice-versa when a
video stops after 9 minutes the system waits another 10 to suspend rather than
suspending within the next minute.

Since this patch has the potential for unwanted side-effects (screen waking up
when download finishes, I could imagine) I'll leave this as 5.4 material and
not part of the 5.3.x patch releases to give it more thorough testing.

BUG: 315438
FIXED-IN: 5.4.0