Load DPMSAction only on xcb platform
The DPMS action is implemented in a very x11 specific way calling into
xcb in the ctor and before supported can be checked. This causes a crash
on platforms != xcb.

In order to better support other platforms the action is marked as
xcb specific in the desktop file and the loading of Actions is adjusted
to only load Actions supported on the current platform.

REVIEW: 124318