Add ConsoleKit2 as a power backend
Internally, it uses the login1 interface with ConsoleKit2 since they share identical DBUS calls.
It also changes the setupSystemdInhibition to setupSeatManagerInhibition so that the login1 and
ConsoleKit2 inhibit calls to minimize code duplication.

This patch doesn't modify the build system to keep it small. A future patch can be added to use
ConsoleKit2 on BSD systems as well.

ConsoleKit2's suspend/hibernate/inhibit DBUS API is documented at:
http://consolekit2.github.io/ConsoleKit2/#Manager

REVIEW: 124388