Add support for toggling screen on/off with a button
This will eventually be used for the phone's power button.
Also, DPMS action needs to be split into a XRandR implementation and a Wayland/KWin one.

On X DPMS just turns on the screen automatically on any input, so pressing the power button
will first cause X to turn it on and then PowerDevil gets the event and turns the screen
right back off, which makes it rather useless for X but KWin won't do that on the phone.