Fix brightness key being ignored during brightness animation
m_cachedBrightnessMap is not updated during animation. When the
animation is running, brightness() function will return the endValue
instead of real hardware value. Checking m_cachedBrightnessMap against
brightness() return value does not make sense. If the animation is
running, skip this check and brightnessKeyPressed will calculate a new
brightness value based on animation's endValue. So the key press will
not be ignored.

REVIEW: 125182