Wait 5s before enforcing an inhibition
Whenever Chrome plays a sound it posts an inhibition. This is great, except that it does that
also when playing a short sound, eg. when receiving a message.
This patch makes PowerDevil wait 5 seconds before actually enforcing the inhibition. In
any case I don't want to have the system wake up for any such short inhibitions.

Also cleanup; the inhibition and inhibition with explicit dbus service methods were
virtually identical.

REVIEW: 126145
BUG: 352235
FIXED-IN: 5.6.0