Revert "Enable animated brightness change also for Sysfs helper"
This reverts commit e3275e994d9b4e2b1676d90a07faed49a445792f.

It seems to trigger a bug in KAuth and more importantly spawns a helper a dozen
times during the animation; the helper should probably do the animation.
This needs some refactoring so the respective backend (kauth / xrandr) do the
animation rather than the already complex upower backend class. Let's not
risk anything for 5.6 and look into it properly at some point.

Somehow this commit didn't end up in Plasma/5.6