Make NetworkManagerQt and BluezQt optional
Summary:
While it is most certainly sensible and helpful on a system equipped
with Bluetooth and Wi-Fi hardware, it serves no purpose on a desktop
system without these and just pulls in a long chain of unwanted
dependencies.

Test Plan: Successfully built with BluezQt and NetworkManagerQt present and missing.

Reviewers: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D1944