[PolicyAgent] Ignore "Keep screen on" request while screen is locked
When an application is blocking power management, such as a video player,
the screen will stay on and will not automatically lock, which is good.

However, if the screen is manually locked it will stay on indefinitely
(until the inhibition is released) which is pointless as you cannot see
the video anymore and you can't do anything with the computer anyway.

When the screen is unlocked, the original policy comes into effect again.

Differential Revision: https://phabricator.kde.org/D1901