Add information about existence of discrete GPU
Summary: This patch adds a kauth helper class which just checks existence of /sys/kernel/debug/vgaswitcheroo/switch file,
which means the user has dual gpu in case of existence. The value we get using that helper class is exposed over DBus so
it can be used outside as well.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3451