rename wireless power saving to wireless
Summary:
This option is too easily confused with Wireless power saving of 801.11x
adapters, where you can turn the feature "wireless power saving" on and
off. I enabled this option, and expected wireless power saving on my
wifi card to go on and off, not my wifi to be disabled.

Since this label is in context of energy saving anyway, simplify the
label to make it more clear.

Screenshot: http://imgur.com/a/SBYcd

Test Plan: Verified the new label is there

Reviewers: #plasma, davidedmundson, broulik

Reviewed By: #plasma, davidedmundson, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4524