skip the disabled backlight device
While adjusting the brightness, if there are more than one graphics card
on the machine, then there might be more than one backlight interface
directories under /sys/backlight. Only one of them is active, and in the
code, KDE picks the first one in the list as the backlight device. It
might pick the wrong one if the active device is the second one.
So, we try skipping the disabled one and can fix the brightness issue on
some machines with more than one graphics cards.

Changes V2:
	1. replace tab by 8 space
	2. add file.close() after the file contain has been read

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>