Remove code to poll the udev events for brightness
Summary:
This code watches the device in the sysfs for changes, if device is
changed it reads the brightness again and updates it.

However that is,

a) Redundant: as if we update the brightness we already know what
brightness we set it to

b) Covers corner case: If something other then powerdevil updates the
brightness, this will update the internal brightness code accordingly

So, lets remove this extra overhead of going through udev/uevent and
make things bit more simpler, Also remove the workaround added for leds
subsystem in the code which was required on Plasma Mobile, Nexus 5(x).
Since we are updating cache value ourself everytime without waiting for
udev and that workaround was doing same.

Test Plan: tested brightness control using keyboard, battery monitor

Reviewers: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5976