[KCM & UI] Use the word "Sleep" instead of "Suspend"
Summary:
[copy-pasted from T10500:]

Problems with the current term:
- It's a somewhat severe-sounding technical term that doesn't intrinsically mean much to the average person. To the extent that it does, the connotations may be negative (e.g. when a misbehaving kid is being kept out of school)
- Our two major closed-source desktop competitors (macOS and Windows) both use the term "Sleep", so user familiarity with "Suspend" is low and falling all the time

I'd like to propose that we use "Sleep" instead, to address the above issues. It's a nicer
word in general, and since it's what our competitors use, we benefit from user familiarity.
We also recently adopted a moon icon with Zs for this, which visually connotes sleep. So it
would make sense to update the string too, because then the string and the icon will both
match.

Test Plan:
Energy Saving KCM: {F6626041}
Advanced [Energy] Settings KCM: {F6626043}

Reviewers: #vdg, #plasma, broulik, filipf, mart

Reviewed By: #vdg, #plasma, filipf, mart

Subscribers: filipf, mart, ltoscano, abetts, aacid, plasma-devel, kde-doc-english

Tags: #plasma, #documentation

Maniphest Tasks: T10500

Differential Revision: https://phabricator.kde.org/D19186