Fixed initial size and content width of Activity Settings KCM
Summary:
Tab contents are added when KActivities::Consumer::serviceStatusChanged signal is fired which happens after the layout is set. This causes the window to be too small initially so a minimumSize is set manually.

This change also adjusts the column span to 2 for rows with a single checkbox. The contents of third column would be pushed outside the scroll area on rows where column 1 is the label and column 2 are the controls since the width of checkboxes in column 1 is too wide.

BUG: 398793
FIXED-IN: 5.16.0

Reviewers: #plasma, broulik, ngraham

Reviewed By: ngraham

Subscribers: ngraham, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20055