Add a global shortcut action to turn off the screen
PowerDevilDPMSAction already provides implementations for X11 and Wayland which allow turning off the screen.
For X11 there was a way to do this on the command line (with "xset dpms force off"), but there's no such replacement for Wayland yet.

Add an action - to which users can assign a shortcut - to turn off the screen.

Differential Revision: https://phabricator.kde.org/D22261