Improve the behaviour of the power button on mobile
Summary:
On Mobile, if the power button is pressed for short time, suspend to Ram as
requested by Bhushan. If the power button is pressed for over a second show a
menu asking whether to reboot, shut down and others like we show on the
desktop.

Albeit pretty cool and useful, this can't be translated into the desktop
because there press time is always 0ms for some reason.

See https://phabricator.kde.org/D26899

Reviewers: #plasma, bshah, broulik

Reviewed By: #plasma, broulik

Subscribers: z3ntu, broulik, PureTryOut, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27193