Add setting for charge threshold
On supported hardware, when using the device mostly on line power, this allows to stop charging
the battery when it reaches a certain level as to not leave the battery fully charge for extended
periods of time. It also allows to configure below which level the battery will then be charged again.

Unfortunately, UPower doesn't seem to be aware of this, so the device will just sit around the
configured percentage, still claiming "charging".

The battery stop charging threshold is exposed on DBus so a hint coul be displayed in Battery
monitor when a limit is configured.