Change "adaptor" to "adapter"
This always seemed like a typo to me. However upon doing some research,
apparently it's a regional spelling difference:
- https://writingexplained.org/adapter-vs-adaptor-difference

It seems that "adaptor" is more common in British English, but still
less common than "adapter". "Adaptor" is virtually unknown in American
English, and KDE uses American English by default, so let's change it to
use that. If "adaptor" is deemed to be appropriate for British English,
it can be used in the en_GB translation.