Watch DBus service right away to discard pending inhibitions reliably
PowerDevil delays applying a power management inhibition by 5 seconds to avoid brief inhibitions
to e.g. wake up the screen and similar.
However, when an inhibition was posted and the process then quit, we would then still
enforce the inhibition as it wasn't explicitly revoked.
Fix this by always watching the inhibition sender and remove it from pending when the
service goes away.

BUG: 423131
FIXED-IN: 5.18.6