[core] introduce the method to wake system up
When using aggressive power management on devices like mobile or laptop
devices, There's generally one objective, saving as much as power as
possible. To do this on some devices we keep system in deep
sleep/suspend for most of the time.

This means userspace will be frozen as well. For most part this is fine,
but for services like push notification helper, or update manager or for
example alarm application, this is not acceptable and it would mean that
we need to wake system at some point.

This helpers on org.kde.Solid.PowerManagement are supposed to help with
that.

You have two possibilities,

- Wake system instantly: this is useful when screen is turned off, but
  device is not in sleep, and you want to show e.g notifications on
  screen.
- Wake system in future: You want to schedule alarm or for example
  wakeup 15 mins later to check for notifications.

Once system is awake, powerdevil will ping dbus service of your choice
and call the method "wakeupCallback". By default this wakeup will be
silent as-in the DPMS and screen backlight will stay off, if you wish to
wake system fully, you can call "wakeup" method to turn system on fully.

Related: #2