Add a series of kconfigxt modules for PowerDevil Sorry about the size of this commit, but it needs to be atomic (I can't change just parts of it, code wouldn't compile)
The main settings module is PowerDevilProfileSettings,
That keeps all of the configuration in place, and all
the groups for a profile. A profile can be "AC", "Battery"
and "LowBattery".

The PowerDevilCurrentProfile stores just that - what profile
is set currently for the system (we need a different setting
for that because the PowerDevilProfileSetting needs to know
which one to load).

PowerDevilActivitySettings is for per-activity profile settings

The PowerDevil core had a massive, and quite weak, code that
loaded the correct profile based on assumptions and saved data
The original code was quite easy to break and unfortunately
the new one is too. I'v tried to keep as much as the old behavior
as possible but test with care.

- About powerdevilprofilegenerator

The biggest issue with this code was the hardcoded
values for possible ways to initialize the profiles.

The KConfigXT xmls currently have <code> sessions for default
values that takes in consideration the profile name and
the default values.

it's still a bit fragile as writting c++ code on a xml is
really not ergonomic, but it appears to work.

- Port the KCM to KConfigXT

This also changes how the KCM loads the widgets. Prior to this
all the widgets where provided by a library from the ActionConfig
plugin, but that will not work correctly for QML.

The new interface is entirely made with a designer xml file,
and much of the handwritten logic could be removed.