Add a series of kconfigxt modules for PowerDevil
The main settings module is PowerDevilProfileSettings,
That keeps all of the configuration in place, and all
the groups for a profile. A profile can be "AC", "Battery"
and "LowBattery".

The PowerDevilCurrentProfile stores just that - what profile
is set currently for the system (we need a different setting
for that because the PowerDevilProfileSetting needs to know
which one to load).

PowerDevilActivitySettings is for per-activity profile settings

The next commit in the series will start to implement the
usecases for them.