Remove all ActionConfig plugins
ActionConfig classes are classes that represents how the
plugin will create the interface, creating the interface
elements manually.

There are three things with this approach that I particularly
dislike:

1 - It's Qt Widgets specific, so it's impossible to use on Qml

2 - It's really hard to visualize the interface on the WYSIWYG
environment such as designer

3 - It's impossible to understand the whole interface as
things might not load on your computer because your computer
does not implement a specific hardware component (and thus,
impossible to verify if the interface is working properly)

4 - It's not deterministic and the interface will vary from
user to user.

A better approach is to implement everything on the designer
and enable / disable things as needed by querying the plugins.

even if it uses a bit more memory for that, it's negligible.