Port PowerDevilCore to the generated KConfigXT Classes
The PowerDevil core had a massive, and quite weak, code that
loaded the correct profile based on assumptions and saved data
The original code was quite easy to break and unfortunately
the new one is too. I'v tried to keep as much as the old behavior
as possible but test with care.