Clear powerdevilprofilegenerator
The biggest issue with this code was the hardcoded
values for possible ways to initialize the profiles.

The KConfigXT xmls currently have <code> sessions for default
values that takes in consideration the profile name and
the default values.

it's still a bit fragile as writting c++ code on a xml is
really not ergonomic, but it appears to work.