Port the KCM to KConfigXT
This also changes how the KCM loads the widgets. Prior to this
all the widgets where provided by a library from the ActionConfig
plugin, but that will not work correctly for QML.

The new interface is entirely made with a designer xml file,
and much of the handwritten logic could be removed.