Use KWin's tablet mode to decide on the mobility UX
There are certain aspects where we decide to use slightly different
settings for a tablety device (e.g. Plasma Mobile) and the rest. So far
this was getting set using the QT_QUICK_CONTROLS_MOBILE environment
variable, which makes it harder for us as it carries tons of graphical
connotations that we don't care for in here. Instead use KWin's
TabletMode, which fits much better what we are trying to do here.

This applies to:
- the power button configuration (on tablet it puts the device to sleep,
on non-tablet it shows the logout page)
- profile generation. Since profile generation is static, we'll use the
availability of the feature to decide how to generate it. This way the
device won't just start behaving completely different when it stops
acting as a tablet.