Fix profile switching in the brightness actions
The current code assumes that the action is notified of every profile change,
which is only the case when the action is configured to run in all profiles.
The code that prevents brightness from being raised when switching to a more
conservative profile gets confused if the action didn't run in the "previous"
profile. Make it more robust by always querying the core for the previous
profile, which it luckily hasn't updated yet when the action gets executed.

BUG: 394945
FIXED-IN: 5.25