Port KCMs to new plasma install namespaces, embed json metadata
Because we have desktop files for the plugins, I have not done the desktop to json
conversion as part of that commit. Once the plugin system stuff is figured out, the
changes should be done in one go, then we can copy the translations and avoid extra work for the translators.

Task https://phabricator.kde.org/T14517