daemon: Fix wakeup removal not rescheduling the timerfd wakeup BUG: 464251
Fixes wakeup removal for the current event from sometimes not rescheduling the timerfd wakeup for the next event, causing the current event to trigger for another cookie.

This happens because error checking condition was invalid; events can get removed from the middle of the list as well.

We experienced this issue in kclock, which can be read about here: https://bugs.kde.org/show_bug.cgi?id=464251