Introduce powerdevilenums.h, replacing two synchronized enum types
SuspendSession as well as ProfileGenerator both had an enum
called Mode (or Modes) with the same enum names and values.

This commit takes PowerDevilEnums from Tomaz Canabrava's unmerged
MR !71, with slight tweaks for better naming, and replaces both of
the original enum declarations and uses with the new
PowerDevil::PowerButtonAction.

In addition to a central place for shared enum declarations,
this class allows for later QML integration via Q_ENUM_NS() macro.