Introduce ProfileDefaults class and use it in ProfileGenerator
ProfileDefaults is meant to be used by the upcoming KConfigXT port of
profile configurations, as a more declarative way to query defaults.
But we also want it to stay in sync with ProfileGenerator as long as
that class still exists.

By using ProfileDefaults in ProfileGenerator::generateProfiles(),
we can both ensure that any change to default values will be
picked up by the future KConfigXT code, while also simplifying
the generateProfiles() function itself.

This commit does not introduce functional changes to profile
generation, care has been taken to ensure that no default values
were changed while being factored out into ProfileDefaults.

The naming of the new default value getters partially matches that
of the existing config groups and keys, but deviates in some places
with an eye on future (hopefully better) KConfigXT naming.