Add a non-singleton version of PowerDevilSettings.kcfgc
The kcmutils_generate_module_data() CMake command doesn't play well
with singleton config classes. Once the global config KCM gets ported
to QML, it would be nice to use this helper command. It's easy to do
without the autogenerated module data class. However, after porting
everything to QML, there won't be much benefit in having the
singleton either.

So in preparation of a porting the global config KCM to QML, this
commit switches the existing QWidgets-based KCM to use a new
non-singleton version of PowerDevilSettings, aptly named
PowerDevil::GlobalSettings.

Plus a minor bit of cleanup to remove includes and KConfig compilation
where we don't actually use the settings class in the first place.