Test ProfileGenerator, so it doesn't change inadvertently
This commit includes a helper tool called generate_profiles, which
writes the generated powermanagementprofilesrc to a custom path for
a given combination of isMobile/canSupportToRam/canSupportToDisk.

Six generated profiles for the most commonly expected options,
generated at this current revision, are included as reference for
a CMake-based autotest that uses generate_profiles and diff.

Future commits that change ProfileGenerator will have to adapt
the reference files if generated files are changed, for CTest to
continue passing.