Hardcode brightness animation duration/threshold
PowerDevilSettings includes brightness animation duration and
threshold settings, which are used in PowerDevilUPowerBackend.
Those settings are not exposed via KCM, but could have been
manually set by the user in the powerdevilrc config file.

In a later commit, we want to parameterize the global settings
object and stop relying on profile generation. But the backend
isn't supposed to have full knowledge of parameterization flags
that's available to PowerDevil::Core.

This commit removes functionality to read the backend-specific
brightness settings from the config file, and instead hardcodes
the default values in PowerDevilUPowerBackend directly.
This gives us flexibility to change PowerDevilSettings for
sole use in PowerDevil::Core and the global config KCM later on.

Any manually added config entries will remain in powerdevilrc,
so we could always decide to bring back configurability of
brightness animation parameters if anyone actually misses this.