daemon: Drop ActionConfig and all of its subclasses
Now that the config UI was rewritten in QML, these QWidgets-based
config UI lego pieces are not in use anymore and won't be needed.

For additional action configuration UI, one needs to either add
to the monolithic QML profiles config KCM, or make a different KCM
altogether that also writes to (parts of) ProfileSettings.